Make Value fully constructible
Previous implementation of the `Value` enum used a hand-crafted `Either`
enum in order to represent values that can either be null (the `Left`
variant) or non-null value (the `Right`) variant.

The problem is that this enum was not exposed in the public API. As
such, it was impossible for `clickhouse_rs` users to build
`Value::Nullable(...)` values.

This commit fixes the problem by removing the `Either` datatype from the
codebase and using the [`either`] crate instead, hence making every
variant of `Value` constructible!

[`either`]: https://docs.rs/either/1.6.1/either/index.html